requirements.txt
fastapi
uvicorn[standard]
redis
python-jose[cryptography]
passlib[bcrypt]
python-multipart
authlib
python-dotenv
email-validator
flet
httpx 
python-dotenv
email-validator